set up basic scroller
